webservice/pyproject.toml
[tool.poetry]
name = "webservice"
version = "0.1.0"
description = "Simple example webservice for fetching quotes"
authors = ["Ruurd Moelker"]
readme = "README.md"

[tool.poetry.dependencies]
python = "3.12.1"
Django = "^5.0.6"
djangorestframework = "^3.15.1"
Markdown = "^3.6"
djangorestframework-yaml = "^2.0.0"
djangorestframework-xml = "^2.0.0"
requests = "^2.31.0"
django-extensions = "^3.2.3"


[tool.poetry.group.dev.dependencies]
django-extensions = "^3.2.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "webservice.settings"


[tool.poetry.scripts]
start = "scripts:run_server"
test = "scripts:test"
